requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
pytest==8.3.4
pytest-django==4.9.0
pytest-pythonpath==0.7.4
PyJWT==2.6.0
requests==2.28.1
Pillow
